# Conflicts:
#	src/main/java/at/ac/htlsteyr/tetris/Model/Game.java

+++ src/main/java/at/ac/htlsteyr/tetris/Saves/JSONhandler.java
@@ -5,6 +5,7 @@ import at.ac.htlsteyr.tetris.Model.Player;
 import com.google.gson.*;
 
 import java.io.File;
+import java.io.FileNotFoundException;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.nio.file.Path;
@@ -23,15 +24,9 @@ public class JSONhandler {
         gson = new GsonBuilder().setPrettyPrinting().create();
     }
 
-    // TODO: Check for duplicate and if so, change value instead of append
-
     public void writeToJSON(String name, int highscore) {
         try {
             Player player = new Player(name, highscore);
-
-            JsonObject jsonObject = new JsonObject();
-            jsonObject.addProperty("name", player.name());
-            jsonObject.addProperty("highscore", player.highscore());
 
             Scanner scanner = new Scanner(jsonFile);
             StringBuilder sb = new StringBuilder();
@@ -39,19 +34,92 @@ public class JSONhandler {
                 sb.append(scanner.nextLine());
             }
 
-            JsonArray jsonArray = gson.fromJson(sb.toString(), JsonArray.class);
-            jsonArray.add(jsonObject);
+            if (checkIfAlreadyInJSON(sb.toString(), name)) {
+                JsonArray jsonArray = gson.fromJson(sb.toString(), JsonArray.class);
+                for (JsonElement element : jsonArray) {
+                    JsonObject object = element.getAsJsonObject();
+                    if (object.get("name").getAsString().equals(name)) {
+                        object.addProperty("highscore", highscore);
+                        break;
+                    }
+                }
 
-            String json = gson.toJson(jsonArray);
-            FileWriter fW = new FileWriter(jsonFile);
-            fW.write(json);
-            fW.close();
+                String json = gson.toJson(jsonArray);
+                FileWriter fW = new FileWriter(jsonFile);
+                fW.write(json);
+                fW.close();
 
-            // Log action
-            System.out.println("Wrote " + player + " to player.json!");
+                // Log action
+                System.out.println("Changed " + player.getName() + " in player.json!");
+            } else {
+                JsonObject jsonObject = new JsonObject();
+                jsonObject.addProperty("name", player.getName());
+                jsonObject.addProperty("highscore", player.getHighscore());
+
+                JsonArray jsonArray = gson.fromJson(sb.toString(), JsonArray.class);
+                jsonArray.add(jsonObject);
+
+                String json = gson.toJson(jsonArray);
+                FileWriter fW = new FileWriter(jsonFile);
+                fW.write(json);
+                fW.close();
+
+                // Log action
+                System.out.println("Wrote " + player + " to player.json!");
+            }
         } catch (IOException e) {
             throw new RuntimeException(e);
         }
+    }
+
+    public void deletePlayerFromJSON(String name) throws IOException {
+        Scanner scanner = new Scanner(jsonFile);
+        StringBuilder sb = new StringBuilder();
+        while (scanner.hasNextLine()) {
+            sb.append(scanner.nextLine());
+        }
+
+        int index = -1;
+        JsonArray array = gson.fromJson(sb.toString(), JsonArray.class);
+        for (int i = 0; i < array.size(); i++) {
+            JsonObject object = array.get(i).getAsJsonObject();
+            if (object.get("name").getAsString().equals(name)) {
+                index = i;
+                break;
+            }
+        }
+
+        if (index != -1) {
+            array.remove(index);
+        }
+
+        String json = gson.toJson(array);
+        FileWriter fW = new FileWriter(jsonFile);
+        fW.write(json);
+        fW.close();
+    }
+
+    private boolean checkIfAlreadyInJSON(String jsonString, String name) {
+        return jsonString.contains(name);
+    }
+
+    public void checkIfJSONisValid() throws IOException {
+        try {
+            Scanner scanner = new Scanner(jsonFile);
+            if (!scanner.hasNextLine()) {
+                addJSONArray(jsonFile);
+            }
+        } catch (FileNotFoundException e) {
+            File f = new File(PATH_TO_JSON.toUri());
+            addJSONArray(f);
+        }
+
+    }
+
+    private void addJSONArray(File file) throws IOException {
+        FileWriter fileWriter = new FileWriter(file);
+        fileWriter.write("[]");
+        fileWriter.close();
     }
 
     public Player getPlayerInfos(String playerName) {
@@ -64,7 +132,6 @@ public class JSONhandler {
 
             JsonArray jsonArray = gson.fromJson(sb.toString(), JsonArray.class);
 
-            System.out.println(jsonArray);
             for (int i = 0; i < jsonArray.size(); i++) {
                 JsonElement jsonElement = jsonArray.get(i);
                 JsonObject jsonObject = jsonElement.getAsJsonObject();
